Drop --game cli switch:
I tried to standardize the switch - make it case insensitive as in:

@@ -117,2 +117,5 @@ def arg(group, dashed, descr, dest, action='store', default=None):
                         dest='gameName',
+                        choices=['oblivion', 'skyrim',
+                                 'Skyrim Special Edition'.lower(), 'fallout4'],
+                        type=str.lower, # lowercase the argument passed
                         help='Specifies the game Wrye Bash should try to '

but then which names should we allow (note the pure ugliness of
'Skyrim Special Edition'.lower(), too late to go back) ? We should of
course use the `game` fsName (or displayName), however barg is
much higher in the import order than bush#setGame - not to mention we
had to document those, for each new game. Then it dawned me:
There is no "game", there is _a path to a game installation_. So -g
dropped, -o is the real one. Note how the docs could not quite explain
what was going on.

This patch also addresses the game picker dialog displaying the fsName
instead of the display name - see:
http://forums.bethsoft.com/topic/1615914-114/#entry25312250
(and a similar post by Arthmoor for fallout3/NV) and cleans up bush
detection algorithm even a bit more.

Mopy/bash/game/__init__.py: badly out of date.